+++ app/src/main/java/net/bible/service/common/ABStringUtils.kt
@@ -37,4 +37,11 @@ object ABStringUtils : StringUtils() {
         }
         return true
     }
+
+    fun toSuperscript(anyString: String): String {
+        val superscripts = charArrayOf('⁰', '¹', '²', '³', '⁴', '⁵', '⁶', '⁷', '⁸', '⁹')
+        return anyString.map { char ->
+            if (char in '0'..'9') superscripts[char - '0'] else char
+        }.joinToString("")
+    }
 }
